+++ zenml/steps/deployer/__init__.py
@@ -12,11 +12,10 @@
 #  or implied. See the License for the specific language governing
 #  permissions and limitations under the License.
 
-from zenml.core.steps.deployer.base_deployer import BaseDeployerStep
+from zenml.steps.deployer.base_deployer import BaseDeployerStep
+from zenml.steps.deployer.gcaip_deployer import GCAIPDeployer
 
 try:
-    from zenml.core.steps.deployer.cortex_deployer import CortexDeployer
+    from zenml.steps.deployer.cortex_deployer import CortexDeployer
 except ModuleNotFoundError:
     pass
-
-from zenml.core.steps.deployer.gcaip_deployer import GCAIPDeployer
