+++ zenml/steps/deployer/base_deployer.py
@@ -17,7 +17,7 @@ from typing import Type
 from tfx.dsl.components.base import executor_spec
 from tfx.dsl.components.base.base_executor import BaseExecutor
 
-from zenml.core.steps import BaseStep
+from zenml.steps import BaseStep
 
 
 class BaseDeployerStep(BaseStep):
